Data Memory
Found the index by taking the read address and using 5 bit decoder to find the index.

For looped through read data and copied bits from data memory if mem read is a flag.

For looped through mem data and copied bits from write data if mem write is a flag.